recognition/application/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-extensions==2.2.9
djangorestframework==3.11.0
idna==2.9
imutils==0.5.3
joblib==0.14.1
numpy==1.18.4
opencv-python==4.2.0.34
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
pika==1.1.0